Update 10-load-data.Rmd
Updated filter code to filter out particles with f_width of greater than 6.0 and less than 0.125mm. Also added a filter for fiber_width that filters out particles with a fiber_width greater than 3.0, which appear to represent primarily fibers or plant materials based on my review of the photos. I have commented this part out so we can use it as needed later.